+++ query/query_test.go
@@ -5869,6 +5869,8 @@ var maxPendingCh chan uint64
 func TestMain(m *testing.M) {
 	x.Init(true)
 
+	odch = make(chan *protos.OracleDelta, 100)
+	maxPendingCh = make(chan uint64, 100)
 	StartDummyZero()
 
 	dir, err := ioutil.TempDir("", "storetest_")
@@ -5909,8 +5911,6 @@ func TestMain(m *testing.M) {
 	err = schema.ParseBytes([]byte(schemaStr), 1)
 	x.Check(err)
 
-	odch = make(chan *protos.OracleDelta, 100)
-	maxPendingCh = make(chan uint64, 100)
 	go updateMaxPending()
 	r := m.Run()
 
@@ -7726,5 +7726,5 @@ func TestExpandAll(t *testing.T) {
 	}
 	`
 	js := processToFastJSON(t, query)
-	require.JSONEq(t, `{"data":{"q":[{"survival_rate":98.990000,"_xid_":"mich","graduation":"1932-01-01T00:00:00Z","path":[{"name":"Glenn Rhee"},{"name":"Andrea"}],"sword_present":"true","friend":[{"name":"Rick Grimes"},{"name":"Glenn Rhee"},{"name":"Daryl Dixon"},{"name":"Andrea"}],"full_name":"Michonne's large name for hashing","follow":[{"name":"Glenn Rhee"},{"name":"Andrea"}],"power":13.250000,"loc":{"type":"Point","coordinates":[1.1,2]},"name":"Michonne","bin_data":"YmluLWRhdGE=","dob_day":"1910-01-01T00:00:00Z","dob":"1910-01-01T00:00:00Z","son":[{"name":"Andre"},{"name":"Helmut"}],"age":38,"school":[{"name":"School A"}],"alive":true,"gender":"female","noindex_name":"Michonne's name not indexed","address":"31, 32 street, Jupiter"}]}}`, js)
+	require.JSONEq(t, `{"data":{"q":[{"~friend":[{"name":"Rick Grimes"}],"survival_rate":98.990000,"_xid_":"mich","graduation":"1932-01-01T00:00:00Z","path":[{"name":"Glenn Rhee"},{"name":"Andrea"}],"sword_present":"true","friend":[{"name":"Rick Grimes"},{"name":"Glenn Rhee"},{"name":"Daryl Dixon"},{"name":"Andrea"}],"full_name":"Michonne's large name for hashing","follow":[{"name":"Glenn Rhee"},{"name":"Andrea"}],"power":13.250000,"loc":{"type":"Point","coordinates":[1.1,2]},"name":"Michonne","bin_data":"YmluLWRhdGE=","dob_day":"1910-01-01T00:00:00Z","dob":"1910-01-01T00:00:00Z","son":[{"name":"Andre"},{"name":"Helmut"}],"age":38,"school":[{"name":"School A"}],"alive":true,"gender":"female","noindex_name":"Michonne's name not indexed","address":"31, 32 street, Jupiter"}]}}`, js)
 }